server/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
sqlalchemy==1.4.46
pydantic==1.10.7
python-jose==3.3.0
passlib[bcrypt]==1.7.4
mysqlclient==2.1.1
alembic==1.10.4